feat(runtime): packed string ABI, async wait, OTLP/TLS, GPU tree reduction, incremental HNSW and real ONNX inference
- strings cross the ABI as NUL-terminated packed UTF-8 (1 byte/address
  unit): allocators, bounded scans, char_at, print/join paths, panic
  message resolution; ManualMemory counters now record real payload bytes
- removed the unused mark-and-sweep GC; manual allocator is the single path
- new spectra.async.task.wait parks on the reactor instead of busy-polling;
  block_on reuses it
- math.abs is checked (E2902) and random_int uses Lemire's unbiased method
  over xorshift64* with zero-seed guard
- OTLP exporter speaks https:// via rustls + webpki-roots (no tokio),
  keeping hand-rolled HTTP/1.1 POST, retries and drop stats
- GPU sum lowered by two-phase workgroup tree reduction; readback only of
  the final scalar (validated on a live adapter vs CPU reference)
- HNSW inserts incrementally (descendant search + neighbor heuristic +
  bidirectional links); persistence v2 with v1 upgrade; recall@10=1.0 on
  2k vectors
- ONNX: ort 2.0.0-rc.13 behind an  feature — session_from_bytes/run/
  session_free hosts, real weights in all export templates, true session
  metadata replaces template op counting

+++ runtime/src/stdlib/math_io.rs
@@ -421,15 +421,15 @@ extern "C" fn std_io_print(ctx: *mut SpectraHostCallContext) -> i32 {
             let value = args[i * 2 + 1];
             let ok = match tag {
                 PRINT_TAG_STR => {
-                    // String buffer stores each byte as a separate i64 slot
-                    let ptr = value as *const i64;
+                    // String buffer: packed UTF-8 bytes, NUL-terminated
+                    let ptr = value as *const u8;
                     if ptr.is_null() {
                         write!(stdout, "(null)").is_ok()
                     } else {
                         let mut bytes: Vec<u8> = Vec::new();
                         let mut offset = 0usize;
                         loop {
-                            let b = *ptr.add(offset) as u8;
+                            let b = *ptr.add(offset);
                             if b == 0 {
                                 break;
                             }
@@ -508,15 +508,15 @@ extern "C" fn std_io_eprint(ctx: *mut SpectraHostCallContext) -> i32 {
             let value = args[i * 2 + 1];
             let ok = match tag {
                 PRINT_TAG_STR => {
-                    // String buffer stores each byte as a separate i64 slot
-                    let ptr = value as *const i64;
+                    // String buffer: packed UTF-8 bytes, NUL-terminated
+                    let ptr = value as *const u8;
                     if ptr.is_null() {
                         write!(stderr, "(null)").is_ok()
                     } else {
                         let mut bytes: Vec<u8> = Vec::new();
                         let mut offset = 0usize;
                         loop {
-                            let b = *ptr.add(offset) as u8;
+                            let b = *ptr.add(offset);
                             if b == 0 {
                                 break;
                             }
